Fix bug in UpdateIsVnfMenu
Fix the unboxing bug reported by Sonar.

Issue-ID: SDC-3112

+++ asdctool/src/main/java/org/openecomp/sdc/asdctool/main/UpdateIsVnfMenu.java
@@ -71,8 +71,8 @@ public class UpdateIsVnfMenu {
 
 			if (updatePropertyOnServiceAtLeastCertified == null) {
 				System.exit(2);
-			} else if (updatePropertyOnServiceAtLeastCertified.intValue() >= 0) {
-				log.debug("Number of updated services is {}",updatePropertyOnServiceAtLeastCertified.intValue());
+			} else if (updatePropertyOnServiceAtLeastCertified >= 0) {
+				log.debug("Number of updated services is {}",updatePropertyOnServiceAtLeastCertified);
 				System.exit(0);
 			}
